feat: Add Twitter logo to footer
This commit adds the Twitter logo to the footer component in order to provide a link to the Zen Browser's Twitter page. This enhances the social media presence of the browser and allows users to easily access the Twitter account.

Refactor the code in the footer component to include the TwitterLogoIcon from the @radix-ui/react-icons package and add a link to the Zen Browser's Twitter page.

Closes #215